修改 类型

+++ golang/ToolGood/Words/StringSearch.go
@@ -93,7 +93,7 @@ func (this *StringSearch) SetKeywords(keywords []string) {
 
 		for key, val := range oldNode.M_values {
 			var index = val.Index
-			newNode.Add(key, allNode2[index])
+			newNode.Add(int32(key), allNode2[index])
 		}
 		for j := 0; j < len(oldNode.Results); j++ {
 			newNode.SetResults(oldNode.Results[j])
@@ -101,9 +101,9 @@ func (this *StringSearch) SetKeywords(keywords []string) {
 
 		if oldNode.Failure != root {
 			for key, val := range oldNode.Failure.M_values {
-				if newNode.HasKey(key) == false {
+				if newNode.HasKey(int32(key)) == false {
 					var index = val.Index
-					newNode.Add(key, allNode2[index])
+					newNode.Add(int32(key), allNode2[index])
 				}
 			}
 			for j := 0; j < len(oldNode.Failure.Results); j++ {
@@ -127,7 +127,7 @@ func (this *StringSearch) FindFirst(text string) string {
 			tn = this._first[t]
 		} else {
 			var b bool
-			b, tn = ptr.TryGetValue(int(int32(t)))
+			b, tn = ptr.TryGetValue(t)
 			if b == false {
 				tn = this._first[t]
 			}
@@ -151,7 +151,7 @@ func (this *StringSearch) FindAll(text string) []string {
 			tn = this._first[t]
 		} else {
 			var b bool
-			b, tn = ptr.TryGetValue(int(int32(t)))
+			b, tn = ptr.TryGetValue(t)
 			if b == false {
 				tn = this._first[t]
 			}
@@ -176,7 +176,7 @@ func (this *StringSearch) ContainsAny(text string) bool {
 			tn = this._first[t]
 		} else {
 			var b bool
-			b, tn = ptr.TryGetValue(int(int32(t)))
+			b, tn = ptr.TryGetValue(t)
 			if b == false {
 				tn = this._first[t]
 			}
@@ -201,7 +201,7 @@ func (this *StringSearch) Replace(text string, replaceChar int32) string {
 			tn = this._first[t]
 		} else {
 			var b bool
-			b, tn = ptr.TryGetValue(int(int32(t)))
+			b, tn = ptr.TryGetValue(t)
 			if b == false {
 				tn = this._first[t]
 			}
